requirements.txt
aiogram == 3.8.0

telegram~=0.0.1
python-dotenv~=1.0.1